fix(verify): keep cascade under 1000 lines; drop unverifiable GmailOauth claim (ARN-213)
Split cascade diagnostics/tests into submodules for the readability ratchet.
Remove ordering(...) fixture invariant that was Unverifiable under ADR-0178
fail-closed and broke reference-spec verification.
